% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFunctions.R
\name{CliffSite}
\alias{CliffSite}
\title{InternalFunctions}
\usage{
CliffSite(
  x,
  introns = NULL,
  direction = "start",
  minreads = 100,
  min.TSS.percent = 5,
  min.local.percent = 1,
  p.cutoff = 0.01,
  window = 6,
  RefTSS = NULL,
  greedy = FALSE,
  plots = 0
)
}
\description{
Internal function for gmmc4
}
