% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFunctions.R
\name{CliffSite2}
\alias{CliffSite2}
\title{InternalFunctions}
\usage{
CliffSite2(
  x,
  SS,
  direction = "start",
  minreads = 100,
  max.degradation.length = NULL,
  min.Nodegradation.length = NULL
)
}
\description{
Internal functions
}
