% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFunctions.R
\name{DTSS}
\alias{DTSS}
\title{InternalFunctions}
\usage{
DTSS(x, groupBy, groups, group1, ratio1, ratio2, method, gs, deltaPSI = 0)
}
\description{
Internal function for FindMarkers

TSS count from GMM

Differentially expressed TSS
}
