% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSSPlot.R
\name{FeaturePlot}
\alias{FeaturePlot}
\title{FeaturePlot}
\usage{
FeaturePlot(
  object,
  assay = "psi",
  cells = NULL,
  TSSs = NULL,
  reduction = "umap",
  dim = 1:2,
  logScale = FALSE,
  groupBy = NULL,
  groups = NULL,
  plot = TRUE,
  shapeForGroup = FALSE,
  point.size = 1,
  order = TRUE,
  plot.theme = NULL,
  base_size = 15,
  color.high = "#642594",
  color.mid = NULL,
  color.low = NULL,
  midpoint = NULL,
  color.na = "grey90",
  ...
)
}
\arguments{
\item{object}{An scATSDataSet from TSS.}

\item{assay}{Which assay to pull expression data from?}

\item{cells, TSSs}{Which cells or TSSs for plot.}

\item{reduction}{Which reduction result are used for plot.}

\item{dim}{Which dimension of reduction result are used for plot.}

\item{groupBy}{The column name of group in scATSDataSet object meta.data.}

\item{groups}{Only for given groups.}

\item{plot}{Plot or return data frame.}

\item{shapeForGroup}{Given shape for group or not?}

\item{point.size, plot.theme, color.low, color.high, color.na, midpoint, color.mid}{Parameters for ggplot.}

\item{order}{Order the point by value or not?}

\item{...}{Additional parameters for \code{scale_colour_gradient2}}
}
\description{
FeaturePlot
}
\concept{visualization}
