% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSS.R
\name{FindMarkers}
\alias{FindMarkers}
\title{FindMarkers}
\usage{
FindMarkers(
  object,
  gene = NULL,
  groupBy,
  group1 = NULL,
  group2 = NULL,
  ratio1 = 0.1,
  ratio2 = 0.1,
  deltaPSI = 0.1,
  majorOnly = TRUE,
  method = NULL,
  cores = 1
)
}
\arguments{
\item{object}{An scATSDataSet from TSS.}

\item{gene}{genes for FindMarkers.}

\item{groupBy}{The column name of group in scATSDataSet object meta.data.}

\item{group1, group2}{Only for given group.}

\item{ratio1, ratio2}{Cut off of cells expressed given TSS.}

\item{deltaPSI}{The minimum delta PSI between compared groups.}

\item{majorOnly}{Whether only the predominant TSS over all cells of each gene is considered, or not.}

\item{method}{Method for statistical test. Currently, only wald.test, prop.test and wilcox.test are supported.}
}
\description{
Find markers of given group
}
