% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFunctions.R
\name{gmmc}
\alias{gmmc}
\title{InternalFunctions}
\usage{
gmmc(
  object,
  region,
  bams,
  gtfrange,
  UTR,
  EBG,
  IBG,
  UTROnly = FALSE,
  greedy = TRUE,
  MinGeneReads = 1000,
  min.TSS.percent = 5,
  min.local.percent = 1,
  p.cutoff = 0.01,
  window = 10,
  scDR = FALSE,
  MinscTSSReads = 5,
  mapqFilter = 255,
  isSecondaryAlignment = FALSE,
  isSupplementaryAlignment = FALSE,
  isDuplicate = FALSE,
  sep = "_"
)
}
\description{
Internal function for TSSCDF
}
