% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFunctions.R
\name{loadbamgene}
\alias{loadbamgene}
\title{InternalFunctions}
\usage{
loadbamgene(
  bam,
  gene,
  region = NULL,
  txdb = NULL,
  exonByGene = NULL,
  exonIntersect = FALSE,
  junctionIntersect = FALSE,
  ignore.strand = FALSE,
  isSupplementaryAlignment = FALSE,
  isSecondaryAlignment = FALSE,
  isDuplicate = NA,
  mapqFilter = 255,
  sep = "_"
)
}
\description{
Internal functions
}
