% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFunctions.R
\name{loadbamregion}
\alias{loadbamregion}
\title{InternalFunctions}
\usage{
loadbamregion(
  bam,
  region,
  txdb = NULL,
  exonIntersect = FALSE,
  junctionIntersect = FALSE,
  isSupplementaryAlignment = FALSE,
  isSecondaryAlignment = FALSE,
  isDuplicate = NA,
  mapqFilter = 255,
  ignore.strand = FALSE,
  sep = "_"
)
}
\description{
Internal functions
}
