% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFunctions.R
\name{loadregionR1}
\alias{loadregionR1}
\title{InternalFunctions}
\usage{
loadregionR1(
  bam,
  region,
  isDuplicate = NA,
  mapqFilter = 255,
  ignore.strand = FALSE,
  sep = "_"
)
}
\description{
Internal functions
}
