% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalFunctions.R
\name{rss}
\alias{rss}
\title{InternalFunctions}
\usage{
rss(
  region,
  bam,
  EBG,
  ExonicReadsOnly = FALSE,
  mapqFilter = 255,
  isSecondaryAlignment = FALSE,
  isSupplementaryAlignment = FALSE,
  isDuplicate = NA,
  sep = "_"
)
}
\description{
Internal function for TSSPlot, read bam reads and calculate start site and coverage for
}
