% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSSSashimiPlot.R
\name{Sashimi}
\alias{Sashimi}
\alias{StartSite}
\title{Sashimi}
\usage{
Sashimi(
  object,
  bam,
  sep = "_",
  gtf,
  txdb = NULL,
  gene = NULL,
  Cells = NULL,
  transcripts = NULL,
  xlimit = NULL,
  MinDepth = 20,
  CellBarcodeName = NULL,
  groupBy = NULL,
  groups = NULL,
  groupLevels = NULL,
  TSS = NULL,
  exonIntersect = TRUE,
  junctionIntersect = FALSE,
  ignore.strand = FALSE,
  mapqFilter = 255,
  isSecondaryAlignment = FALSE,
  isSupplementaryAlignment = FALSE,
  isDuplicate = NA,
  fill.color = NULL,
  line.type = NULL,
  line.color = NULL,
  ColourTxName = "Red",
  base_size = 15,
  rel_height = NULL,
  free_y = TRUE,
  Group.name.angle = NULL
)

StartSite(
  object,
  bam,
  sep = "_",
  gtf,
  txdb = NULL,
  gene = NULL,
  Cells = NULL,
  transcripts = NULL,
  xlimit = NULL,
  MinDepth = 20,
  CellBarcodeName = NULL,
  groupBy = NULL,
  groups = NULL,
  groupLevels = NULL,
  TSS = NULL,
  exonIntersect = TRUE,
  junctionIntersect = FALSE,
  ignore.strand = FALSE,
  mapqFilter = 255,
  isSecondaryAlignment = FALSE,
  isSupplementaryAlignment = FALSE,
  isDuplicate = NA,
  point.color = "#FEB24C",
  point.size = 1,
  point.shape = 16,
  line.type = NULL,
  line.color = NULL,
  ColourTxName = "Red",
  base_size = 15,
  rel_height = NULL,
  free_y = FALSE,
  Group.name.angle = NULL
)
}
\arguments{
\item{object}{An scATSDataSet object.}

\item{gtf}{The file path pf GTF file, or GRanges of GTF file imported by readGFFAsGRanges of rtracklayer package.}

\item{gene}{The target gene.}

\item{Cells}{Which cells will by remained.}

\item{transcripts}{Which transcripts will by remained in the reference panel.}

\item{groupBy, groupLevels}{Group the cells, and which groups will be used for Sashimi.}

\item{TSS}{The loci of TSSs to be indicated in in Sashimi.}

\item{fill.color}{The color(s) of Sashimi of each group.}

\item{line.type, line.color}{The line type and line colour of indicated TSS.}

\item{base_size}{base font size of ggplot theme, given in pts.}

\item{rel_height}{relative height of reference panel.}

\item{free_y}{Free the y dimension of grouped Sashimi?}

\item{reads}{The mapped reads, imported by loadbam function.}

\item{which}{A GRanges, or any object that can be coerced to a GRanges, or missing object.}

\item{adjust, bw}{the smoothing bandwidth to be used of density ridges.}
}
\description{
A function for TSS Sashimi plot.
}
\concept{visualization}
