% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSS.R
\name{ThetaByGroup}
\alias{ThetaByGroup}
\title{ThetaByGroup}
\usage{
ThetaByGroup(object, gene, cells = NULL, groupBy, group1 = NULL, group2 = NULL)
}
\arguments{
\item{object}{An scATSDataSet from TSS.}

\item{gene}{the name of gene.}

\item{cells}{Which cells to use.}

\item{groupBy}{The column name of group in scATSDataSet object meta.data.}

\item{group1, group2}{Only for given group.}
}
\description{
ThetaByGroup
}
