% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSS.R
\name{TSSCDF}
\alias{TSSCDF}
\title{TSS inference and quantification}
\usage{
TSSCDF(
  object,
  bam,
  sep = "_",
  genes,
  gtf,
  txdb = NULL,
  UTROnly = FALSE,
  MinGeneReads = 500,
  mapqFilter = 255,
  isSecondaryAlignment = FALSE,
  isSupplementaryAlignment = FALSE,
  isDuplicate = NA,
  min.TSS.percent = 5,
  min.local.percent = 1,
  p.cutoff = 0.01,
  window = 10,
  greedy = TRUE,
  cores = 1L,
  verbose = FALSE,
  project.name = NULL
)
}
\arguments{
\item{object}{A Seurat object.}

\item{bam, sep}{Path of bam file(s), and the character string to separate the bams.}

\item{genes}{Genes for ATS inference and quantification.}

\item{gtf}{Path of GTF file or GRanges file of GTF file.}

\item{txdb}{A \code{TxDb} object.}

\item{UTROnly}{Only infer ATS within annotated 5'-UTR regions or first exons of transcripts without 5'-UTR.}

\item{MinGeneReads}{The minimum reads of a gene for TSS inference.}

\item{mapqFilter, isSecondaryAlignment, isSupplementaryAlignment, isDuplicate}{parameters for \code{ScanBamParam} to read bam file(s).}

\item{min.TSS.percent}{The minimum percent of a TSS.}

\item{min.local.percent}{The minimum difference between observed and local average derivative of CDF of a TSS.}

\item{p.cutoff}{The cutoff of p-value of a TSS.}

\item{window}{The window for local mean value estimating.}

\item{greedy}{greedy for the first TSS? If this parameter is set to TRUE, the first TSS will be printed even if it does not satisfy the specified condition.}

\item{cores}{The number of cores for parallel working.}

\item{verbose}{A logical controlling if a text progress bar is displayed.}

\item{project.name}{The name of project.}
}
\description{
TSS inference and quantification
}
