% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSSPlot.R
\name{TSSPlot}
\alias{TSSPlot}
\title{TSSPlot}
\usage{
TSSPlot(
  object,
  bam,
  gtf,
  txdb = NULL,
  gene = NULL,
  Cells = NULL,
  ExonicReadsOnly = FALSE,
  CellBarcodeName = NULL,
  sep = "_",
  MaxSigma = Inf,
  TSS = NULL,
  MinProp = 0,
  groupBy = NULL,
  groups = NULL,
  groupLevels = NULL,
  adjust = 1,
  mapqFilter = 255,
  isSecondaryAlignment = FALSE,
  isSupplementaryAlignment = FALSE,
  isDuplicate = FALSE,
  binwidth = 25,
  base_size = 15,
  FillSashimi = "#FEB24C",
  ColourTxName = "Red",
  rel_heights = c(1, 1, 5, 1)
)
}
\arguments{
\item{object}{An scATSDataSet from TSS.}

\item{bam}{Path of bam file.}

\item{gtf}{Path of GTF file or GRanges file of GTF file.}

\item{gene}{Gene for plot.}

\item{ExonicReadsOnly}{Only the R2 of reads  within annotated exon region are remained or not.}

\item{CellBarcodeName}{The column name of cell barcode (CB) in Seurat object meta.data.}

\item{sep}{The character string to separate the CB (cell barcode) and sample name.}

\item{MaxSigma, MinProp, TSS}{Limitation of inferred TSS for plotting.}

\item{groupBy}{The column name of group in scATSDataSet object meta.data.}

\item{groups, groupLevels}{Only for given groups.}

\item{adjust}{adjust for density ridges plot.}

\item{mapqFilter}{The minimum map quqlity.}

\item{isSupplementaryAlignment, isSecondaryAlignment}{parameters for ScanBamParam.}

\item{binwidth}{binwidth for histogram plot.}

\item{base_size, FillSashimi, ColourTxName}{Plot parameters.}

\item{rel_heights}{Relative heights of final plots.}
}
\description{
Find markers of given group
}
\concept{visualization}
